+++ next.config.js
@@ -1,3 +1,7 @@
 const withPlugins = require("next-compose-plugins");
 
-module.exports = withPlugins([], {});
+module.exports = withPlugins([], {
+    images: {
+        domains: ["s3.us-west-2.amazonaws.com"],
+    },
+});