+++ src/util/router.ts
@@ -11,3 +11,9 @@ export const isActiveLink = (
 
     return currentPathname.startsWith(href);
 };
+
+export const getCanonicalURL = (title: string): string => {
+    const cleaned = title.replace(/[^-\]_.~!*'();:@&=+$,/?%#[A-z0-9]/g, " ");
+    const removedSpaces = cleaned.split(" ").join("-");
+    return removedSpaces;
+};